+++ Presentation/OpenTgResearcherDesktop/ViewModels/TgChatDetailsParticipantsViewModel.cs
@@ -1,7 +1,8 @@
-﻿namespace OpenTgResearcherDesktop.ViewModels;
+﻿
+namespace OpenTgResearcherDesktop.ViewModels;
 
 [DebuggerDisplay("{ToDebugString()}")]
-public sealed partial class TgChatDetailsParticipantsViewModel : TgPageViewModelBase
+public sealed partial class TgChatDetailsParticipantsViewModel : TgPageViewModelBase, IDisposable
 {
     #region Fields, properties, constructor
 
@@ -19,6 +20,12 @@ public sealed partial class TgChatDetailsParticipantsViewModel : TgPageViewModel
     [ObservableProperty]
     public partial Action ScrollRequested { get; set; } = () => { };
 
+    /// <summary> Cancellation token for download session </summary>
+    private CancellationTokenSource? _loadCts;
+    private CancellationToken _loadToken;
+
+    public IAsyncRelayCommand LoadParticipantsCommand { get; }
+    public IAsyncRelayCommand StopParticipantsCommand { get; }
     public IAsyncRelayCommand GetParticipantsCommand { get; }
 
     public TgChatDetailsParticipantsViewModel(ITgSettingsService settingsService, INavigationService navigationService, ILogger<TgChatDetailsParticipantsViewModel> logger,
@@ -27,7 +34,58 @@ public sealed partial class TgChatDetailsParticipantsViewModel : TgPageViewModel
     {
         AppNotificationService = appNotificationService;
         // Commands
+        LoadParticipantsCommand = new AsyncRelayCommand(LoadParticipantsAsync);
+        StopParticipantsCommand = new AsyncRelayCommand(StopParticipantsAsync);
         GetParticipantsCommand = new AsyncRelayCommand(GetParticipantsAsync);
+    }
+
+    #endregion
+
+    #region IDisposable
+
+    /// <summary> To detect redundant calls </summary>
+    private bool _disposed;
+
+    /// <summary> Finalizer </summary>
+	~TgChatDetailsParticipantsViewModel() => Dispose(false);
+
+    /// <summary> Throw exception if disposed </summary>
+    public void CheckIfDisposed() => ObjectDisposedException.ThrowIf(_disposed, this);
+
+    /// <summary> Release managed resources </summary>
+    public void ReleaseManagedResources()
+    {
+        _loadCts?.Cancel();
+        _loadCts?.Dispose();
+        _loadToken = CancellationToken.None;
+    }
+
+    /// <summary> Release unmanaged resources </summary>
+    public void ReleaseUnmanagedResources()
+    {
+        //
+    }
+
+    /// <summary> Dispose pattern </summary>
+    public void Dispose()
+    {
+        // Dispose of unmanaged resources
+        Dispose(true);
+        // Suppress finalization
+        GC.SuppressFinalize(this);
+    }
+
+    /// <summary> Dispose pattern </summary>
+    private void Dispose(bool disposing)
+    {
+        if (_disposed) return;
+        // Release managed resources
+        if (disposing)
+            ReleaseManagedResources();
+        // Release unmanaged resources
+        ReleaseUnmanagedResources();
+        // Flag
+        _disposed = true;
     }
 
     #endregion
@@ -37,7 +95,8 @@ public sealed partial class TgChatDetailsParticipantsViewModel : TgPageViewModel
     public override async Task OnNavigatedToAsync(NavigationEventArgs? e) => await LoadStorageDataAsync(async () =>
     {
         Uid = e?.Parameter is Guid uid ? uid : Guid.Empty;
-        await LoadDataStorageCoreAsync();
+        //await LoadParticipantsCoreAsync();
+        await Task.CompletedTask;
     });
 
     protected override async Task SetDisplaySensitiveAsync()
@@ -50,16 +109,44 @@ public sealed partial class TgChatDetailsParticipantsViewModel : TgPageViewModel
         await Task.CompletedTask;
     }
 
-    private async Task LoadDataStorageCoreAsync()
+    private async Task LoadParticipantsAsync() => await ContentDialogAsync(LoadParticipantsCoreAsync, TgResourceExtensions.AskLoading(), TgEnumLoadDesktopType.Storage);
+
+    private async Task LoadParticipantsCoreAsync()
     {
         try
         {
+            _loadCts?.Cancel();
+            _loadCts?.Dispose();
+            _loadCts = new CancellationTokenSource();
+            _loadToken = _loadCts.Token;
+
             if (!SettingsService.IsExistsAppStorage) return;
 
-            Dto = await App.BusinessLogicManager.StorageManager.SourceRepository.GetDtoAsync(x => x.Uid == Uid);
-            
+            Dto = await App.BusinessLogicManager.StorageManager.SourceRepository.GetDtoAsync(x => x.Uid == Uid, _loadToken);
+            await Task.Delay(250);
+
+            // Single query to get all distinct users for this source
+            UserDtos = [.. await App.BusinessLogicManager.StorageManager.UserRepository.GetUsersBySourceIdAsync(Dto.Id, _loadToken)];
+
             IsEmptyData = false;
             ScrollRequested?.Invoke();
+        }
+        finally
+        {
+            await ReloadUiAsync();
+            await SetDisplaySensitiveAsync();
+        }
+    }
+
+    private async Task StopParticipantsAsync() => await ContentDialogAsync(StopParticipantsCoreAsync, TgResourceExtensions.AskStopLoading(), TgEnumLoadDesktopType.Storage);
+
+    private async Task StopParticipantsCoreAsync()
+    {
+        try
+        {
+            _loadCts?.Cancel();
+            _loadCts?.Dispose();
+            _loadToken = CancellationToken.None;
         }
         finally
         {