+++ Presentation/OpenTgResearcherDesktop/ViewModels/TgStoriesViewModel.cs
@@ -43,7 +43,7 @@ public sealed partial class TgStoriesViewModel : TgPageViewModelBase
 
     private void ClearDataStorageCore() => Dtos.Clear();
 
-    private async Task LoadDataStorageAsync() => await ContentDialogAsync(LoadDataStorageCoreAsync, TgResourceExtensions.AskDataLoad(), TgEnumLoadDesktopType.Storage);
+    private async Task LoadDataStorageAsync() => await ContentDialogAsync(LoadDataStorageCoreAsync, TgResourceExtensions.AskLoading(), TgEnumLoadDesktopType.Storage);
 
 	private async Task LoadDataStorageCoreAsync()
 	{